Move out common methods
